Changed float to int of Experience feat. Changed from drop to give to player. Drop XP on break machine.

+++ forge/src/main/java/es/degrassi/forge/init/entity/type/IExperienceEntity.java
@@ -1,5 +1,5 @@
 package es.degrassi.forge.init.entity.type;
 
 public interface IExperienceEntity extends IDegrassiEntity {
-  void setXp(float xp);
+  void setXp(int xp);
 }
